requirements.txt
# Core dependencies
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1

# LangChain and AI (using latest versions with thought_signature support)
langchain-core>=1.2.0
langchain-google-genai>=4.2.0
langchain>=1.2.0
langchain-chroma>=1.1.0
langchain-community>=0.4.0
langchain-text-splitters>=1.1.0
langgraph>=1.0.0

# Data validation
pydantic>=2.10.6,<3.0.0

# Vector database
chromadb>=0.5.23

# Utilities
google-genai>=1.0.0
